discord-bot/requirements.txt
discord.py
python-dotenv
google-auth
requests